+++ q.js
@@ -39,13 +39,12 @@ try {
 var qStartingLine = captureLine();
 var qFileName;
 
-// shims
+require("es5-shim");
+var asap = require("asap");
+var WeakMap = require("weak-map");
 
 // used for fallback in "allResolved"
 var noop = function () {};
-
-require("es5-shim");
-var asap = require("asap");
 
 var object_create = Object.create || function (prototype) {
     function Type() { }
@@ -171,8 +170,9 @@ function deprecate(callback, name, alternative) {
     };
 }
 
-// end of shims
-// beginning of real work
+// end of long stack traces
+
+var handlers = new WeakMap();
 
 /**
  * Constructs a promise for an immediate reference, passes promises through, or
@@ -195,6 +195,32 @@ function Q(value) {
     }
 }
 Q.resolve = Q;
+
+/**
+ * TODO
+ */
+function getHandler(promise) {
+    var handler = handlers.get(promise);
+    if (!handler || !handler.became) {
+        return handler;
+    }
+    handler = followHandler(handler);
+    handlers.set(promise, handler);
+    return handler;
+}
+
+function followHandler(handler) {
+    if (!handler.became) {
+        return handler;
+    } else {
+        handler.became = followHandler(handler.became);
+        return handler.became;
+    }
+}
+
+var theViciousCycleError = new Error("Can't resolve a promise with itself");
+var theViciousCycleRejection = reject(theViciousCycleError);
+var theViciousCycleHandler = getHandler(theViciousCycleRejection);
 
 /**
  * Performs a task in a future turn of the event loop.
@@ -226,43 +252,24 @@ function defer() {
     // forward to the resolved promise.  We coerce the resolution value to a
     // promise using the `resolve` function because it handles both fully
     // non-thenable values and other thenables gracefully.
-    var messages = [], progressListeners = [], resolvedPromise;
+    var messages = [], progressListeners = [];
 
     var deferred = object_create(defer.prototype);
     var promise = object_create(Promise.prototype);
 
-    promise.promiseDispatch = function (resolve, op, operands) {
-        var args = Array.prototype.slice.call(arguments);
-        if (messages) {
-            messages.push(args);
+    var handler = {
+        dispatch: function (resolve, op, operands) {
+            messages.push([resolve, op, operands]);
             if (op === "when" && operands[1]) { // progress operand
                 progressListeners.push(operands[1]);
             }
-        } else {
-            asap(function () {
-                resolvedPromise.promiseDispatch.apply(resolvedPromise, args);
-            });
-        }
-    };
-
-    // XXX deprecated
-    promise.valueOf = deprecate(function () {
-        if (messages) {
-            return promise;
-        }
-        var nearerValue = nearer(resolvedPromise);
-        if (isPromise(nearerValue)) {
-            resolvedPromise = nearerValue; // shorten chain
-        }
-        return nearerValue;
-    }, "valueOf", "inspect");
-
-    promise.inspect = function () {
-        if (!resolvedPromise) {
+        },
+        inspect: function () {
             return { state: "pending" };
         }
-        return resolvedPromise.inspect();
     };
+
+    handlers.set(promise, handler);
 
     if (Q.longStackSupport && hasStacks) {
         try {
@@ -278,17 +285,22 @@ function defer() {
         }
     }
 
-    // NOTE: we do the checks for `resolvedPromise` in each method, instead of
-    // consolidating them into `become`, since otherwise we'd create new
-    // promises with the lines `become(whatever(value))`. See e.g. GH-252.
-
     function become(newPromise) {
-        resolvedPromise = newPromise;
+        handler.became = theViciousCycleHandler;
+        var newHandler = getHandler(newPromise);
+        handler.became = newHandler;
+        // TODO eliminate promise.source, use getHandler(promise).became
         promise.source = newPromise;
 
+        handlers.set(promise, newHandler);
+        handler = void 0; // A dead resolver should not retain dead objects
+        promise = void 0;
+
         messages.forEach(function (message) {
+            // TODO figure out whether this asap is necessary.  makeQ does not
+            // have it.
             asap(function () {
-                newPromise.promiseDispatch.apply(newPromise, message);
+                newHandler.dispatch.apply(void 0, message);
             });
         });
 
@@ -297,30 +309,25 @@ function defer() {
     }
 
     deferred.promise = promise;
+
     deferred.resolve = function (value) {
-        if (resolvedPromise) {
+        if (!messages) {
             return;
         }
 
         become(Q(value));
     };
 
-    deferred.fulfill = function (value) {
-        if (resolvedPromise) {
-            return;
-        }
-
-        become(fulfill(value));
-    };
     deferred.reject = function (reason) {
-        if (resolvedPromise) {
+        if (!messages) {
             return;
         }
 
         become(reject(reason));
     };
+
     deferred.notify = function (progress) {
-        if (resolvedPromise) {
+        if (!messages) {
             return;
         }
 
@@ -462,43 +469,33 @@ function Promise(descriptor, fallback, inspect) {
 
     var promise = object_create(Promise.prototype);
 
-    promise.promiseDispatch = function (resolve, op, args) {
-        var result;
-        try {
-            if (descriptor[op]) {
-                result = descriptor[op].apply(promise, args);
-            } else {
-                result = fallback.call(promise, op, args);
+    var handler = {
+        dispatch: function (resolve, op, operands) {
+            var result;
+            try {
+                if (descriptor[op]) {
+                    result = descriptor[op].apply(promise, operands);
+                } else {
+                    result = fallback.call(promise, op, operands);
+                }
+            } catch (exception) {
+                result = reject(exception);
             }
-        } catch (exception) {
-            result = reject(exception);
-        }
-        if (resolve) {
-            resolve(result);
-        }
+            if (resolve) {
+                resolve(result);
+            }
+        },
+        inspect: inspect
     };
 
-    promise.inspect = inspect;
-
-    // XXX deprecated `valueOf` and `exception` support
-    if (inspect) {
-        var inspected = inspect();
-        if (inspected.state === "rejected") {
-            promise.exception = inspected.reason;
-        }
-
-        promise.valueOf = deprecate(function () {
-            var inspected = inspect();
-            if (inspected.state === "pending" ||
-                inspected.state === "rejected") {
-                return promise;
-            }
-            return inspected.value;
-        });
-    }
+    handlers.set(promise, handler);
 
     return promise;
 }
+
+Promise.prototype.inspect = function () {
+    return getHandler(this).inspect();
+};
 
 Promise.prototype.toString = function () {
     return "[object Promise]";
@@ -535,7 +532,7 @@ Promise.prototype.then = function (fulfilled, rejected, progressed) {
     }
 
     asap(function () {
-        self.promiseDispatch(function (value) {
+        getHandler(self).dispatch(function (value) {
             if (done) {
                 return;
             }
@@ -553,7 +550,7 @@ Promise.prototype.then = function (fulfilled, rejected, progressed) {
     });
 
     // Progress propagator need to be attached in the current tick.
-    self.promiseDispatch(void 0, "when", [void 0, function (value) {
+    getHandler(this).dispatch(void 0, "when", [void 0, function (value) {
         var newValue;
         var threw = false;
         try {
@@ -640,9 +637,7 @@ function nearer(value) {
  */
 Q.isPromise = isPromise;
 function isPromise(object) {
-    return isObject(object) &&
-        typeof object.promiseDispatch === "function" &&
-        typeof object.inspect === "function";
+    return isObject(object) && !!handlers.get(object);
 }
 
 Q.isPromiseAlike = isPromiseAlike;
@@ -656,11 +651,11 @@ function isPromiseAlike(object) {
  */
 Q.isPending = isPending;
 function isPending(object) {
-    return isPromise(object) && object.inspect().state === "pending";
+    return Q(object).isPending();
 }
 
 Promise.prototype.isPending = function () {
-    return this.inspect().state === "pending";
+    return getHandler(this).inspect().state === "pending";
 };
 
 /**
@@ -1004,10 +999,10 @@ function dispatch(object, op, args) {
 }
 
 Promise.prototype.dispatch = function (op, args) {
-    var self = this;
     var deferred = defer();
+    var self = this;
     asap(function () {
-        self.promiseDispatch(deferred.resolve, op, args);
+        getHandler(self).dispatch(deferred.resolve, op, args);
     });
     return deferred.promise;
 };